pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "monotributo-cli"
version = "0.1.0"
description = "CLI MVP para emisión de Factura C (Monotributo) usando PyAfipWs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "click>=8.1",
  "pyafipws",
  "python-dotenv>=1.0",
  "setuptools>=82",
]

[project.scripts]
monofact = "monofact.cli:main"

[dependency-groups]
dev = [
  "pytest>=8.0",
]

[tool.setuptools]
packages = ["monofact"]

[tool.uv.sources]
pyafipws = { path = "../pyafipws", editable = true }
